refactored out SFTP related download/upload into remote_file_storage_wrappers.py and fixed unit tests
