+++ .travis.yml
@@ -12,6 +12,6 @@ before_install:
 - 'if [ "${TRAVIS_PULL_REQUEST}" = "false" ]; then openssl aes-256-cbc -K $encrypted_d17283647768_key -iv $encrypted_d17283647768_iv
   -in test.synapseConfig.enc -out test.synapseConfig -d; mv test.synapseConfig ~/.synapseConfig; openssl aes-256-cbc -K $encrypted_d17283647768_key -iv $encrypted_d17283647768_iv
   -in test.awscredentials.enc -out test.awscredentials -d; mkdir -p ~/.aws && mv test.awscredentials ~/.aws/credentials;fi'
-- travis_wait pip install pysftp cython pandas
+- travis_wait pip install pysftp cython pandas boto3
 addons:
   ssh_known_hosts: ec2-54-212-85-156.us-west-2.compute.amazonaws.com
