+++ synapseclient/utils.py
@@ -641,7 +641,7 @@ def printTransferProgress(transferred, toBeTransferred, prefix = '', postfix='',
 
     """
     if not sys.stdout.isatty():
-        pass#return
+        return
     barLength = 20 # Modify this to change the length of the progress bar
     status = ''
     rate = ''
