+++ tests/unit/unit_test_client.py
@@ -10,7 +10,7 @@ import synapseclient
 from synapseclient import Evaluation, File, concrete_types, Folder
 from synapseclient.exceptions import *
 from synapseclient.dict_object import DictObject
-
+import synapseclient.upload_functions as upload_functions
 
 def setup(module):
     print('\n')
@@ -246,19 +246,25 @@ def test__uploadExternallyStoringProjects_external_user(mock_upload_destination)
     expected_storage_location_id = "1234567"
     expected_local_state = {'_file_handle':{}, 'synapseStore':True}
     expected_path = "~/fake/path/file.txt"
+    expected_path_expanded = os.path.expanduser(expected_path)
+    expected_file_handle_id = "8786"
     mock_upload_destination.return_value = {'storageLocationId' : expected_storage_location_id,
                                             'concreteType' : concrete_types.EXTERNAL_S3_UPLOAD_DESTINATION}
 
     test_file = File(expected_path, parent="syn12345")
 
-    # method under test
-    path, local_state,  storage_location_id = syn._Synapse__uploadExternallyStoringProjects(test_file, local_state={'_file_handle':{}, 'synapseStore':True}) #dotn care about filehandle for this test
 
-    #test
-    mock_upload_destination.assert_called_once_with(test_file)
-    assert_equal(expected_path, path)
-    assert_equal(expected_local_state, local_state)
-    assert_equal(expected_storage_location_id, storage_location_id)
+    # method under test
+    with patch.object(synapseclient.upload_functions, "multipart_upload", return_value=expected_file_handle_id) as mocked_multipart_upload, \
+         patch.object(syn.cache, "add") as mocked_cache_add,\
+         patch.object(syn, "_getFileHandle") as mocked_getFileHandle:
+        file_handle = upload_functions.upload_file(syn, test_file['parentId'], test_file.local_state()) #dotn care about filehandle for this test
+
+        mock_upload_destination.assert_called_once_with(test_file['parentId'])
+        mocked_multipart_upload.assert_called_once_with(syn, expected_path_expanded, contentType=None, storageLocationId=expected_storage_location_id)
+        mocked_cache_add.assert_called_once_with(expected_file_handle_id,expected_path_expanded)
+        mocked_getFileHandle.assert_called_once_with(expected_file_handle_id)
+        #test
 
 def test_login__only_username_config_file_username_mismatch():
     if (sys.version < '3'):
